add DBConnection
